+++ kf_lib/actors/fighter/_moves.py
@@ -13,7 +13,7 @@ class MoveMethods(BaseFighter):
         self.learn_move(random.choice(sample).name)
 
     def get_moves_to_choose(self, tier):
-        return moves.get_rand_moves(self, self.num_moves_choose, tier)
+        return moves.get_rand_moves(self, self.num_moves_choose)
 
     def get_move_tier_for_lv(self):
         return ceil(self.level / 2)
@@ -48,7 +48,6 @@ class MoveMethods(BaseFighter):
             self.learn_move(mv)
             # print(f'{self} learns {mv}')
 
-    # todo in learn_random_move, if move_tier is not given, make it relative to current level
     def learn_random_move(self, move_tier, silent=False):
         move_obj = moves.get_rand_move(self, move_tier)
         self.learn_move(move_obj, silent=silent)
@@ -82,8 +81,7 @@ class MoveMethods(BaseFighter):
     def set_rand_moves(self):
         for lv in LVS_GET_NEW_ADVANCED_MOVE:
             if self.level >= lv:
-                tier = self.get_move_tier_for_lv()
-                self.learn_move(moves.get_rand_moves(self, 1, tier)[0])
+                self.learn_move(moves.get_rand_move(self))
         for lv, moves_to_learn in self.style.move_strings.items():
             if self.level >= lv:
                 if isinstance(moves_to_learn, str):  # can be a tuple, can be a string