+++ docs/version_history.md
@@ -26,7 +26,8 @@ Date
 22. feat: remove backflip strikes
 23. **feat: combinations of move functions for generation are now described in a table ("move_word_combinations.csv"), not in the code** 
 24. feat: new "Drunken" moves, revamp move word combinations (generated 3095 -> 3727 moves)
-25. refactor strike multipliers
+25. refactor: strike multipliers
+26. feat: common logger
 
 ### Coming soon: